// src/apis/models/TenantSettings.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Feature } from './Feature';
import { HttpFile } from '../http/http';

export class TenantSettings {
  'features': Array<Feature>;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'features',
      baseName: 'features',
      type: 'Array<Feature>',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return TenantSettings.attributeTypeMap;
  }

  public constructor() {}
}